# src/openapi.yaml
openapi: 3.0.3
info:
  description: API for systems used by the employees of Zlagoda shops
  title: Zlagoda
  version: 1.0.0
servers:
- url: https://localhost:3333
...
paths:
...
  /api/product-category/:
    get:
      operationId: getProductCategoryList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryCriteria'
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryListResponse'
          description: List of product categories
      summary: Get list of product categories
      tags:
      - product-category
    put:
      operationId: createProductCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: Id of new product category
      summary: Create a new product category
      tags:
      - product-category
  /api/product-category/{id}:
    delete:
      operationId: deleteProductCategory
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether productCategory was deleted
      summary: Delete a product category by id
      tags:
      - product-category
    get:
      operationId: getProductCategoryById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategory'
          description: Product category by id
      summary: Get product category by id
      tags:
      - product-category
    post:
      operationId: updateProductCategory
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether product category was updated
      summary: Update existing product category
      tags:
      - product-category
...
  /api/receipt:
    get:
      operationId: getReceiptList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptCriteria'
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptListResponse'
          description: List of receipts
      summary: Get list of receipts
      tags:
      - receipt
    put:
      operationId: createReceipt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: Id of new receipt
      summary: Create a new receipt
      tags:
      - receipt
  /api/receipt/{id}:
    delete:
      operationId: deleteReceipt
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether receipt was deleted
      summary: Delete a receipt by id
      tags:
      - receipt
    get:
      operationId: getReceiptById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
          description: Receipt by id
      summary: Get receipt by id
      tags:
      - receipt
...
components:
  schemas:
...
    ProductCategoryView:
      example:
        name: name
      properties:
        name:
          title: name
          type: string
      required:
      - name
      title: ProductCategoryView
      type: object
    ProductCategory:
      allOf:
      - properties:
          id:
            type: integer
        required:
        - id
        type: object
      - $ref: '#/components/schemas/ProductCategoryView'
      example:
        name: name
        id: 6
      title: ProductCategory
...
    ReceiptItemView:
      example:
        product: 6
        quantity: 1
      properties:
        product:
          title: product
          type: integer
        quantity:
          title: quantity
          type: integer
      required:
      - product
      - quantity
      title: ReceiptItemView
      type: object
    ReceiptItem:
      example:
        quantity: 9
        price: 3
        productArchetype: 7
      properties:
        productArchetype:
          title: productArchetype
          type: integer
        quantity:
          title: quantity
          type: integer
        price:
          title: price
          type: integer
      required:
      - price
      - productArchetype
      - quantity
      title: ReceiptItem
      type: object
    ReceiptView:
      example:
        customerCardId: 0
        items:
        - product: 6
          quantity: 1
        - product: 6
          quantity: 1
      properties:
        customerCardId:
          title: customerCardId
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ReceiptItemView'
          title: items
          type: array
      required:
      - items
      title: ReceiptView
      type: object
    Receipt:
      example:
        dateTime: 2000-01-23T04:56:07.000+00:00
        totalPrice: 5
        customerCardId: 5
        vat: 2
        cashierId: 1
        id: 6
        items:
        - quantity: 9
          price: 3
          productArchetype: 7
        - quantity: 9
          price: 3
          productArchetype: 7
      properties:
        id:
          title: id
          type: integer
        cashierId:
          title: cashierId
          type: integer
        customerCardId:
          title: customerCardId
          type: integer
        dateTime:
          format: date-time
          title: dateTime
          type: string
        totalPrice:
          title: totalPrice
          type: integer
        vat:
          title: vat
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ReceiptItem'
          title: items
          type: array
      required:
      - cashierId
      - dateTime
      - id
      - items
      - totalPrice
      - vat
      title: Receipt
      type: object
...
    ProductCategoryListResponse:
      example:
        totalCount: 0
        items:
        - name: name
          id: 6
        - name: name
          id: 6
      properties:
        totalCount:
          title: totalCount
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ProductCategory'
          title: items
          type: array
      required:
      - items
      - totalCount
      title: ProductCategoryListResponse
      type: object
...
    ReceiptListResponse:
      example:
        totalCount: 0
        items:
        - dateTime: 2000-01-23T04:56:07.000+00:00
          totalPrice: 5
          customerCardId: 5
          vat: 2
          cashierId: 1
          id: 6
          items:
          - quantity: 9
            price: 3
            productArchetype: 7
          - quantity: 9
            price: 3
            productArchetype: 7
        - dateTime: 2000-01-23T04:56:07.000+00:00
          totalPrice: 5
          customerCardId: 5
          vat: 2
          cashierId: 1
          id: 6
          items:
          - quantity: 9
            price: 3
            productArchetype: 7
          - quantity: 9
            price: 3
            productArchetype: 7
      properties:
        totalCount:
          title: totalCount
          type: integer
        items:
          items:
            $ref: '#/components/schemas/Receipt'
          title: items
          type: array
      required:
      - items
      - totalCount
      title: ReceiptListResponse
      type: object
...
    ProductCategoryCriteria:
      allOf:
      - $ref: '#/components/schemas/BaseCriteria'
      example: {}
      title: ProductCategoryCriteria
...
    ReceiptCriteria:
      allOf:
      - $ref: '#/components/schemas/BaseCriteria'
      - properties:
          cashierIds:
            items:
              type: integer
            type: array
          startDate:
            format: date-time
            type: string
          endDate:
            format: date-time
            type: string
        type: object
      example:
        cashierIds:
        - 0
        - 0
        endDate: 2000-01-23T04:56:07.000+00:00
        startDate: 2000-01-23T04:56:07.000+00:00
      title: ReceiptCriteria
...
    BaseCriteria:
      allOf:
      - properties:
          offset:
            type: integer
          limit:
            type: integer
        type: object
      - $ref: '#/components/schemas/SortCriteria'
      title: BaseCriteria
    SortCriteria:
      properties:
        sortField:
          title: sortField
          type: string
        sortAscending:
          title: sortAscending
          type: boolean
      title: SortCriteria
      type: object